Added method is_conversion_ready() to class INA219
Proposing the method is_conversion_ready() to class INA219 for checking if conversion was done. The method is checking bit CNVR in the voltage register according to INA219 datasheet. This function is useful when streaming data seamlessly and without losses. With this is_conversion_ready() can be used for checking if conversion was done before reading the next measurement results.